Revert "update"
This reverts commit 35605c34ce61acff04c331996a4b2a204d1e6e57.

+++ src/sample/Accueil.java
@@ -16,7 +16,7 @@ import javafx.stage.Stage;
 
 public class Accueil extends Parent {
 
-    public Accueil(Stage primaryStage, double width, double height) {
+    public Accueil(Stage primaryStage, int width, int height) {
 
         //Labels
         Label t_c = new Label("Entrez c :");
